Fix GC checkpoint CI failures
